neatness changes

+++ fillit_dir/zeewier_auxiliary.c
@@ -25,7 +25,7 @@ void			remove_tetri(struct s_tetrimino *tetrimino, uint64_t *grid)
 	grid[((*tetrimino).cubes[4][0] + (*tetrimino).cubes[2][0]) / 4] ^=
 		1ull << (63 - ((*tetrimino).cubes[4][1] + (*tetrimino).cubes[2][1] +
 			(((*tetrimino).cubes[4][0] + (*tetrimino).cubes[2][0]) % 4 * 16)));
-	(*tetrimino).placed = (short int)0;
+	(*tetrimino).placed = 0;
 }
 
 void			place_tetri(struct s_tetrimino *tetrimino, uint64_t *grid)
